\name{clusSummaryBernoulliDiscrete}
\alias{clusSummaryBernoulliDiscrete}
\alias{clusSummaryBinomialNormal}
\alias{clusSummaryCategoricalDiscrete}
\alias{clusSummaryNormalDiscrete}
\alias{clusSummaryNormalNormal}
\alias{clusSummaryPoissonDiscrete}
\alias{clusSummaryPoissonNormal}
\alias{clusSummaryVarSelectBernoulliDiscrete}
\title{Definition of characteristics of sample datasets for profile regression}
\description{.}
\usage{
clusSummaryBernoulliDiscrete()
clusSummaryBernoulliDiscrete()
clusSummaryBinomialNormal()
clusSummaryCategoricalDiscrete()
clusSummaryNormalDiscrete()
clusSummaryNormalNormal()
clusSummaryPoissonDiscrete()
clusSummaryPoissonNormal()
clusSummaryVarSelectBernoulliDiscrete()
}
\value{
Need to write this 
\item{none}{List of stuff}
}
\author{David Hastie and Silvia Liverani}
\examples{

clusSummaryBernoulliDiscrete()

}
\keyword{simulation}

