\name{computeAssociatedVariable}
\alias{computeAssociatedVariable}
\title{computeAssociatedVariable}
\description{computeAssociatedVariable.}
\usage{
computeAssociatedVariable(subjectValues,clusObj,clusterPlotOrder,latexFile=NULL)
}
\arguments{
\item{subjectValues}{dont know}
\item{clusObj}{dont know}
\item{clusterPlotOrder}{dont know}
\item{latexFile}{dont know}
}
\value{
Need to look into this, I don't know how it compares them.
}
\author{David Hastie and Silvia Liverani}



